Update SAR_cfg_grps_lingor.sqf
Check that some of your areas overlap / are in the same area. Survivor camp and desert base are pretty much the same area. Bandit_Camp and El
_Presedente are exactly the same areas. Burnt_House and Bandit_Outpost are again very close.

If you dont have a good reason to do this, combine them to one area.

Another thing:

I set _this setMarkeralpha 1; to 1 for your areas, by this you can see the areas if you are the first to login after a server restart.

After you have finished testing and are fine with everything, set this to _this setMarkeralpha 0; for all areas, and they will not show up anymore.